backend/requirements.txt
# ============================================================
# CONTEXTUS - Voice Assistant Backend Requirements
# ============================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.24.1

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangChain & LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# RAG (Retrieval Augmented Generation)
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.2

# Database & Supabase
supabase==2.3.0
postgrest==0.13.0

# Additional utilities
loguru==0.7.2  # Enhanced logging
